pyproject.toml
[project]
name = "vista-mcp-server"
version = "0.1.0"
description = "MCP server for Vista Enterprise, AP, Project, Vendor, and Health APIs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "mcp",
    "pyjwt>=2.10.0",
    "cryptography>=44.0.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.5.2",
    "celery>=5.6.3",
    "redis>=7.4.0",
]

[project.scripts]
vista-mcp-server = "server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]
